# specs/002-jwt-auth/contracts/auth-openapi.yaml
openapi: 3.1.0
info:
  title: Todo Backend API - Authentication Extension
  description: |
    Authentication endpoints and security schemes for the Todo Backend API.
    Extends the base API with JWT-based authentication.
  version: 2.0.0

servers:
  - url: http://localhost:8000
    description: Local development server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token issued by Better Auth.
        Include in Authorization header as: Bearer <token>

  schemas:
    User:
      type: object
      required:
        - id
        - email
        - created_at
      properties:
        id:
          type: string
          format: uuid
          description: Unique user identifier (from Better Auth)
          example: "550e8400-e29b-41d4-a716-446655440000"
        email:
          type: string
          format: email
          description: User email address
          example: "user@example.com"
        created_at:
          type: string
          format: date-time
          description: When user was first synced
          example: "2026-01-10T12:00:00Z"

    AuthError:
      type: object
      required:
        - detail
      properties:
        detail:
          type: string
          description: Error message
          example: "Not authenticated"

    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Task title (required, non-empty)
          example: "Complete authentication feature"
        description:
          type: string
          nullable: true
          description: Optional task description
          example: "Implement JWT verification in FastAPI"

    TaskUpdate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 500
          description: Updated task title
        description:
          type: string
          nullable: true
          description: Updated task description

    TaskResponse:
      type: object
      required:
        - id
        - user_id
        - title
        - is_completed
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Task ID
          example: 1
        user_id:
          type: string
          description: Owner user ID
          example: "550e8400-e29b-41d4-a716-446655440000"
        title:
          type: string
          example: "Complete authentication feature"
        description:
          type: string
          nullable: true
          example: "Implement JWT verification"
        is_completed:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/AuthError'
          examples:
            not_authenticated:
              summary: Missing token
              value:
                detail: "Not authenticated"
            invalid_token:
              summary: Invalid or expired token
              value:
                detail: "Invalid token"
            expired_token:
              summary: Token has expired
              value:
                detail: "Token expired"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: string
                example: "Task not found"

    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            type: object
            properties:
              detail:
                type: array
                items:
                  type: object
                  properties:
                    loc:
                      type: array
                      items:
                        type: string
                    msg:
                      type: string
                    type:
                      type: string

security:
  - BearerAuth: []

paths:
  /api/tasks:
    get:
      summary: List all tasks for authenticated user
      operationId: listTasks
      tags:
        - Tasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user's tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      summary: Create a new task for authenticated user
      operationId: createTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'

  /api/tasks/{task_id}:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
        description: Task ID

    get:
      summary: Get a specific task by ID
      operationId: getTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      summary: Update an existing task
      operationId: updateTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Task updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'

    delete:
      summary: Delete a task
      operationId: deleteTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Task deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/tasks/{task_id}/complete:
    parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: integer
        description: Task ID

    patch:
      summary: Mark a task as completed
      operationId: completeTask
      tags:
        - Tasks
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Task marked as completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  /api/me:
    get:
      summary: Get current authenticated user info
      operationId: getCurrentUser
      tags:
        - Auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Current user information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'

tags:
  - name: Tasks
    description: Task management endpoints (protected)
  - name: Auth
    description: Authentication status endpoints